Restructure the code for easier apps integration
